Fix NotFound error that can occur during cuncurrent access/deletes
Summary: If a row is found by client A, but then deleted by client B before client A can lock or read it we were generating an internal error with the useless error message of "NotFound".  This diff resolves the problem by detecting the condition and simply moving on to the next row.

Test Plan: MTR

Reviewers: hermanlee4, spetrunia, anthony, yoshinorim

Reviewed By: yoshinorim

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D53205