Delete dead code which isn't part of the build
Summary: Delete code which has been added in 2014, but is not part of the build and isn't used anywhere. The underlying goal is to make sure that the operational codebase reflects reality and is as light as possible. That way code is easier to understand, debug, refactor etc.

Test Plan:
```
mysqlbuild.sh
mysqlbuild.sh --release
```

Reviewers: alexyang, jkedgar

Reviewed By: jkedgar

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D55473