Write primary keys to SstFileWriter in bulk_load mode
Summary: For performance reasons we want to be able to write primary keys directly to SSTs since they are already in order when doing a bulk load.  This diff redirects the primary key portion of the row write to an SstFileWriter and manages handling all the resulting SST files.

Test Plan: Still working on what tests to add

Reviewers: yoshinorim, hermanlee4

Reviewed By: hermanlee4

Subscribers: alexyang, IslamAbdelRahman, sdong, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D59643
Differential Revision: https://reviews.facebook.net/D62109