Fix performance problem caused by incorrect usage of shared_ptr
Summary:
We were passing around references to shared_ptr<Rdb_key_def> and sometimes using shared_ptr<const Rdb_key_def>.  Whenever we passed one into the other a new shared_ptr was temporarily constructed and then destructed.  This is the wrong way to do this.  We should only be passing the shared_ptr object when we want to take ownership.

As is often the case, fully understanding the issue leads to cleaner/simpler code.

Fixes Issue 330 (https://github.com/facebook/mysql-5.6/issues/330)

Test Plan: MTR and ran through the code under the debugger to make sure that a shared_ptr was not being constructed/destructed as before

Reviewers: yoshinorim, spetrunia, MarkCallaghan, hermanlee4

Reviewed By: hermanlee4

Subscribers: mung, vasilep, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D64605