Fix the varbinary encoding to avoid an inefficiency
Summary:
The existing varbinary format has an inefficiency such that any source that has a length of a multiple of 8 bytes will generate an encoded value that has an extra 9 bytes that can be avoided.

The pack/unpack/skip functions needed access to the key's version information, so instead of passing it in as another parameter (which gets ugly), I made all of these functions member functions of Rdb_key_def.

While investigating I realized that for short input values breaking in multiples of 8 bytes causes quite a bit of overhead.  The new scheme breaks the input into 4 byte segments for the first 32 bytes of input and after that it uses 8 byte segments.  This provides a good compromise between the savings that 8 byte segments gives for long inputs and 4 byte segments give for short inputs.

Reviewed By: hermanlee

Differential Revision: D4637146

fbshipit-source-id: 1509321