Fix rocksdb_sys_vars.rocksdb_force_flush_memtable_and_lzero_now_basic
Summary:
When run as part of the RocksDBBig test set, rocksdb_sys_vars.rocksdb_force_flush_memtable_and_lzero_now_basic was failing.  The problem is that if a compaction is already happening on the specified file a kAborted status is returned, but this is interpreted as a failure by the MyRocks code.

Squash with: D4509351

Reviewed By: lth, gunnarku

Differential Revision: D4803800

fbshipit-source-id: c3cde95